remove "Unknow security" Error
When we use pdblp to request data for a list of tickers,  if any one of the tickers is invalid, we get errors and can't data for the rest of the them.    We have to pre-check all tickers one by one to handle the errors.  It's very inefficient.   If we remove this ValueError, it will return nan for invalid tickers which is easy to handle in the dataframe.